+++ .github/release-notes.md
@@ -1,3 +1,4 @@
 * Fixed encounters list overflowing with long names
 * Added Smissmas and Halloween sections to casual selection manager
 * Made all of the casual page scrollable
+* Better detection and handling of server-side bots